% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spLink_occ.r
\name{data_prep}
\alias{data_prep}
\title{Download occurrences from spLink}
\usage{
data_prep(splist, token = NULL)
}
\arguments{
\item{splist}{\code{vector}species binomial (i.e., c("Lepus saxatilis","Puma concolor"))}

\item{token}{\code{vector}personal token obtained in spLink plataform}

\item{api_list}{\code{list} of species API paths obtained from \code{data_prep} function}

\item{lst}{\code{boolen} indicating if output shoul return a list of data frame (lst = TRUE) or a single long data.frame (lst = FALSE)}
}
\value{
\code{data.frame} occurrences obtained from spLink
}
\description{
This function allows you to download occurrence records from the Brazilian spLink database in \code{.csv} format for one or multiple species.
}
\examples{
\dontrun{
spList <- data_prep(    
splist = c("Perna perna","Perna viridis","Abramites hypselonotus","Callithrix jacchus"), 
token = "adicine seu token"
)
opree_spLink(api_list = spList, lst = TRUE)
}
}
