% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{db_collection}
\alias{db_collection}
\alias{db_splist}
\alias{db_oco}
\alias{geometrias}
\title{database collection description}
\source{
Observatorio Paranaense de Especies Exoticas
}
\description{
The main data sources used in opree database
}
\keyword{datasets}
