% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_occ.r
\name{opree_dB}
\alias{opree_dB}
\title{occurrences database}
\usage{
opree_dB(raw_ = TRUE, subset_ = NULL, key_ = NULL)
}
\arguments{
\item{raw_}{\code{boleano} Default is \code{TRUE} return a data.frame with species occurrences}

\item{subset_}{\code{vector} with the condition to be filtered}

\item{key_}{\code{vector} with the variable to the key variable to be filtered}
}
\value{
return data.frame with species name, longitude, latitude and available year of observation
}
\description{
Main file containing occurrence records from the collection of databases
}
\examples{
\dontrun{
opree_dB(raw = FALSE,subset_ = "PELD", key_ = "dB")
}
}
