% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_taxa_id.r
\name{opree_exo_lst}
\alias{opree_exo_lst}
\title{opree taxa assessment}
\usage{
opree_exo_lst(list_lv = TRUE)
}
\arguments{
\item{list_lv}{\code{bolean} determine the level of identification in raw data}
}
\value{
Default return a list of vector containing different levels of identification.
If \code{list_lv = FALSE} return a list of vector with non-valid classificatopm level
}
\description{
This function allow to access the groups of taxa in the raw file of database
}
