% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_splist.r
\name{opree_exo_tbl}
\alias{opree_exo_tbl}
\title{opree species list}
\usage{
opree_exo_tbl(sp_val = TRUE)
}
\arguments{
\item{sp_val}{\code{bolean} determine the level of confidence of species in species list}
}
\value{
Default is \code{TRUE} and return a tibble containing the species with high confidence in their non-native status and prefered environment.
}
\description{
This function allow to access the up-to date non-native species list
}
