% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_eco.r
\name{opree_lifeform}
\alias{opree_lifeform}
\title{Access ecological information from opree}
\usage{
opree_lifeform(clean_ = NULL)
}
\arguments{
\item{clean_}{\code{boolean} If TRUE, return speceis with high confidence of invasisve status. If \code{NULL} retunr all the data}
}
\description{
Access biological information from opree databases
}
