% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_plot_maps.r
\name{opree_map}
\alias{opree_map}
\title{Plot grid maps derived from opree data}
\usage{
opree_map(data_ = NULL, pallete_ = "Spectral", lgd_break = 8, scale_label = "")
}
\arguments{
\item{pallete_}{\code{string} any color sequence derived from \code{RColorBrewer}}

\item{lgd_break}{\code{string} scale of color values. Default is 8}

\item{scale_label}{= "string" Set the title of color range palete}

\item{data_`data.frame`}{generated from \code{opree_spat_grid} function}

\item{name_}{\code{string} plot title}
}
\description{
This function uses occurrence records to estimate the spatial extent of records using grid custom grid cells
}
\examples{
\dontrun{
vt <- db_splist \%>\% filter(eco_evo_class == "terrestrial_vertebrate") \%>\% pull(species)
sp <- db_oco \%>\% filter(species_adj \%in\% vt) 
pr_ <- geometrias \%>\% filter(class == "OpenStreeMap") \%>\% vect()
teste <- opree_spat_grid(
   data_ = sp, 
   shape_ = pr_, 
   long_ = "long_dec", 
   lat_ = "lat_dec",
   area = 100, 
   hex_ = FALSE
)

opree_map(teste, lgd_break = 8)
}
}
