% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_spat_coverage.r
\name{opree_spat_grid}
\alias{opree_spat_grid}
\title{Estimate extent of occurrences using controlled grid cell}
\usage{
opree_spat_grid(
  data_,
  shape_,
  long_,
  lat_,
  area_,
  square_grid = c(TRUE, FALSE)
)
}
\arguments{
\item{data_}{\code{data.frame} Data frame or tible containing longitude and latitude in decimals}

\item{shape_}{\code{polygon} spatial polygon of extent of interest.}

\item{long_}{\code{string} name of longitude variable}

\item{lat_}{\code{string} name of latitude variable}

\item{area_}{\code{numeric} number with the size of ares of interest.}

\item{square_grid}{\code{boolena} to determine if grid should be hexagonal or not}

\item{hex_}{\code{boolean} If \code{TRUE} grid cells in hexagon.}
}
\value{
\code{data.frame} containg the number of occurrences per grid cell
}
\description{
mapping occurrence extent using controled grid cell
}
\examples{
\dontrun{
#' vt <- db_splist \%>\% dplyr::filter(eco_evo_class == "terrestrial_vertebrate") \%>\% dplyr::pull(species)
sp <- db_oco \%>\% dplyr::filter(species_adj \%in\% vt) 
pr_ <- geometrias \%>\% dplyr::filter(class == "OpenStreeMap") \%>\% terra::vect()
grid_plot <- opree_spat_grid(
   data_ = sp, 
   shape_ = pr_, 
   long_ = "long_dec", 
   lat_ = "lat_dec",
   area = 100, 
   hex_ = FALSE
 )
}
}
