% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_spLink_occ.r
\name{opree_spLink}
\alias{opree_spLink}
\title{Download occurrences from spLink}
\usage{
opree_spLink(api_list, lst = c(TRUE, FALSE))
}
\arguments{
\item{api_list}{\code{list} of species API paths obtained from \code{data_prep} function}

\item{lst}{\code{boolen} indicating if output should return a list of data frame (lst = TRUE) or a single long data.frame (lst = FALSE)}
}
\description{
This function allows you to download occurrence records from the Brazilian spLink database in \code{.csv} format for one or multiple species.
}
