% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_collection.r
\name{opree_collection}
\alias{opree_collection}
\title{Database collection}
\usage{
opree_collection(refs_ = NULL)
}
\arguments{
\item{refs_}{\code{bolean} parameter to determine if return references associated with databases}
}
\value{
Default return a vector with database identification code. If refs_ = TRUE return a tibble with database identification code and the references associated.
}
\description{
This function allow to access the sources of databases and their references
}
