% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_cp_idx.r
\name{opree_cp}
\alias{opree_cp}
\title{Vulnerability index}
\usage{
opree_cp(data_, long, lat, shp_, shape_var, data_var, area = 1e+06)
}
\arguments{
\item{data_}{\code{data.frame} with occurrence records, species name or species groups}

\item{long}{\code{string} longitude name in data frame}

\item{lat}{\code{string} latitude name in data frame}

\item{shp_}{\code{spatVec} geometies in \code{spatVec} class}

\item{shape_var}{\code{string} variable of interest inside geometry in \code{shp_}}

\item{data_var}{\code{string} variable name with grupo of interest (i.e. species) em \code{data_}}

\item{area}{\code{numeric} conversion unity scale. Default is 1e6}
}
\value{
\verb{data frame} containing the estimated indicator value, classification rank, and relative percentage for each polygon existing in the \code{geom} parameter
}
\description{
This index uses colonization pressure by kilometer square of spatial area to map the accumulation of non-native species for any spatial file collection
}
\examples{
\dontrun{
vt <- db_splist \%>\% filter(eco_evo_class == "terrestrial_vertebrate") \%>\% pull(especie)
sp <- db_oco \%>\% filter(especie_ajustado \%in\% vt)
municipios <- geometrias \%>\% filter(class == "IAT") \%>\% vect()
ecor <- geometrias \%>\% filter(class == "IBGE") \%>\% vect()
ind <- opree_cp(
   data_ = sp, 
   long = "long_dec", 
   lat = "lat_dec", 
   shp_ = ecor,
   shape_var = "nome", 
   data_var = "especie_ajustado"
)
} 
}
