% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_occ.r
\name{opree_dB}
\alias{opree_dB}
\title{occurrences database}
\usage{
opree_dB(simplify_ = NULL)
}
\arguments{
\item{simplify_}{\code{boleano} If \code{TRUE} return a list of data.frame for each database contributor}
}
\value{
return data.frame with species name, longitude, latitude and available year of observation
}
\description{
Main database of opree containing occurrence records from the collection of databases
}
