% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_occ.r
\name{opree_gbif}
\alias{opree_gbif}
\title{Download occurrences from GBIF}
\usage{
opree_gbif(taxa, login, password, mail, dir)
}
\arguments{
\item{taxa}{\code{vector} speceis binomial (i.e., c("Lepus saxatilis","Puma concolor"))}

\item{login}{\code{string} personal login used to log in GBIF}

\item{password}{\code{string} personal password used to log in GBIF}

\item{mail}{\code{string} personal email used to log in GBIF}

\item{dir}{\code{string} path to store downloaded file}
}
\value{
List of occurrence data and metadata. List element \code{Data} store occurrences in \code{csv} format while \code{Meta_data} store summary metadata including occurrence \code{DOI}.
}
\description{
This function allows you to download GBIF occurrence records in \code{.csv} format for one or multiple species.
The function is adjusted to search for records other than: \verb{uncertainty greater than 1km of precision}, \verb{null geographic coordinates (i.e., 0)}, \verb{invalid coordinates}, \verb{coordinates with error}, \verb{records of dead specimens}, \verb{zoos and botanical gardens}.
In this way, the function returns only records that refer to: \verb{preserved specimens}, \verb{cited material}, \verb{human observation}, \verb{machine records (i.e., photography, filming)}
}
