% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_plot_cp.r
\name{opree_map_cp}
\alias{opree_map_cp}
\title{Plot colonization pressure from \code{opree_cp}}
\usage{
opree_map_cp(data_, shp_, lgd_break, name_, pallete_ = "Spectral")
}
\arguments{
\item{data_}{\verb{data frame ou tibble} with occurrence records and species name}

\item{shp_}{\code{sf} geometry file in sf class}

\item{lgd_break}{\code{string} scale of color values}

\item{name_}{\code{string} plot title}

\item{pallete_}{\code{string} any color sequence derived from \code{RColorBrewer}}

\item{data_`data.frame`}{from \code{opree_cp} function}
}
\description{
This function uses occurrence records to estimate colonization pressure measured by the number of exotic species per square kilometer of the unit of interest.#'
}
\examples{
\dontrun{
vt <- db_splist \%>\% filter(eco_evo_class == "terrestrial_vertebrate") \%>\% dplyr::pull(especie)
sp <- db_oco \%>\% filter(especie_ajustado \%in\% vt) 
ecor <- geometrias \%>\% filter(class == "IBGE") \%>\% vect()
ind <- opree_cp(
data_ = sp, 
long = "long_dec", 
lat = "lat_dec", 
shp_ = ecor,
shape_var = "nome", 
data_var = "especie_ajustado"
)
opree_map_cp(
data_ = ind,
shp = ecor,
lgd_break = 6,
name_ = "teste"
)
}
}
