% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_spat_coverage.r
\name{opree_spat_grid}
\alias{opree_spat_grid}
\title{Estimate extent of occurrences using controlled grid cell}
\usage{
opree_spat_grid(
  data_,
  shape_,
  long_,
  lat_,
  area_,
  square_grid = c(TRUE, FALSE)
)
}
\arguments{
\item{data_}{\code{data.frame} Data frame or tible containing longitude and latitude in decimals}

\item{shape_}{\code{polygon} spatial polygon of extent of interest.}

\item{long_}{\code{string} name of longitude variable}

\item{lat_}{\code{string} name of latitude variable}

\item{area_}{\code{numeric} number with the size of ares of interest.}

\item{hex_}{\code{boolean} If \code{TRUE} grid cells in hexagon.}
}
\value{
\code{data.frame} containg the number of occurrences per grid cell
}
\description{
mapping occurrence extent using controled grid cell
}
